feat: merge code from latest rpa-auth and robot repository.

+++ backend/robot-service/src/main/java/com/iflytek/rpa/base/service/impl/CGlobalVarServiceImpl.java
@@ -6,15 +6,18 @@ import com.iflytek.rpa.base.entity.CGlobalVar;
 import com.iflytek.rpa.base.entity.dto.BaseDto;
 import com.iflytek.rpa.base.entity.dto.CGlobalDto;
 import com.iflytek.rpa.base.service.CGlobalVarService;
-import com.iflytek.rpa.starter.exception.NoLoginException;
-import com.iflytek.rpa.starter.utils.response.AppResponse;
+import com.iflytek.rpa.common.feign.RpaAuthFeign;
+import com.iflytek.rpa.common.feign.entity.User;
 import com.iflytek.rpa.utils.IdWorker;
-import com.iflytek.rpa.utils.UserUtils;
-import java.util.List;
-import javax.annotation.Resource;
+import com.iflytek.rpa.utils.exception.NoLoginException;
+import com.iflytek.rpa.utils.exception.ServiceException;
+import com.iflytek.rpa.utils.response.AppResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import javax.annotation.Resource;
+import java.util.List;
 
 /**
  * 客户端-全局变量(CGlobalVar)表服务实现类
@@ -30,10 +33,18 @@ public class CGlobalVarServiceImpl implements CGlobalVarService {
     @Autowired
     private IdWorker idWorker;
 
+    @Autowired
+    RpaAuthFeign rpaAuthFeign;
+
     @Override
     @RobotVersionAnnotation
     public AppResponse<?> getGlobalVarInfoList(BaseDto baseDto) throws NoLoginException {
-        String userId = UserUtils.nowUserId();
+        AppResponse<User> response = rpaAuthFeign.getLoginUser();
+        if (response == null || !response.ok()) {
+            throw new ServiceException("用户信息获取失败");
+        }
+        User loginUser = response.getData();
+        String userId = loginUser.getId();
         baseDto.setCreatorId(userId);
         List<CGlobalVar> processNameList = globalVarDao.getGlobalVarInfoList(baseDto);
         return AppResponse.success(processNameList);
@@ -42,7 +53,12 @@ public class CGlobalVarServiceImpl implements CGlobalVarService {
     @Override
     @Transactional(rollbackFor = Exception.class)
     public AppResponse<?> createGlobalVar(CGlobalDto globalDto) throws NoLoginException {
-        String userId = UserUtils.nowUserId();
+        AppResponse<User> response = rpaAuthFeign.getLoginUser();
+        if (response == null || !response.ok()) {
+            throw new ServiceException("用户信息获取失败");
+        }
+        User loginUser = response.getData();
+        String userId = loginUser.getId();
         globalDto.setCreatorId(userId);
         globalDto.setUpdaterId(userId);
         int count = globalVarDao.countVarByName(globalDto);
@@ -70,7 +86,12 @@ public class CGlobalVarServiceImpl implements CGlobalVarService {
 
     @Override
     public AppResponse<?> getGlobalVarNameList(String robotId) throws NoLoginException {
-        String userId = UserUtils.nowUserId();
+        AppResponse<User> response = rpaAuthFeign.getLoginUser();
+        if (response == null || !response.ok()) {
+            throw new ServiceException("用户信息获取失败");
+        }
+        User loginUser = response.getData();
+        String userId = loginUser.getId();
         List<CGlobalVar> globalVarNameList = globalVarDao.getGlobalVarNameList(userId, robotId);
         return AppResponse.success(globalVarNameList);
     }
@@ -78,7 +99,12 @@ public class CGlobalVarServiceImpl implements CGlobalVarService {
     @Override
     @Transactional(rollbackFor = Exception.class)
     public AppResponse<?> deleteGlobalVar(CGlobalDto globalDto) throws NoLoginException {
-        String userId = UserUtils.nowUserId();
+        AppResponse<User> response = rpaAuthFeign.getLoginUser();
+        if (response == null || !response.ok()) {
+            throw new ServiceException("用户信息获取失败");
+        }
+        User loginUser = response.getData();
+        String userId = loginUser.getId();
         globalDto.setCreatorId(userId);
         boolean result = globalVarDao.deleteGlobalVar(globalDto);
         return AppResponse.success(result);
